move AlbumImageItem so that it can be reused
